Add utils and general refactor
