+++ src/buk.c
@@ -1,7 +1,7 @@
 #include <stdio.h>		// For printf, fprintf, perror
 #include <stdlib.h>		// For EXIT_SUCCESS, EXIT_FAILURE
 #include <string.h>		// For strcmp
-#include "globals.h"
+#include "utils/globals.h"
 #include "help/help.h"
 #include "init/init.h"
 // #include "config/config.h"
