+++ src/utils/globals.h
@@ -1,0 +1,9 @@
+#ifndef GLOBALS_H
+#define GLOBALS_H
+
+#define NAME "buk"
+#define DIR_NAME "." NAME
+#define CONFIG_PATH "." NAME "/config"
+#define DEFAULT_BACKUP_DIR "." NAME "_backups"
+
+#endif // GLOBALS_H
